% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add two numbers together}
\usage{
add_two_numbers(a, b)
}
\arguments{
\item{a}{(numeric) A positive or negative number}

\item{b}{(numeric) A positive or negative number}
}
\value{
The value of a and b added together
}
\description{
Add two numbers together
}
\examples{
add_two_numbers(a=1,b=4)
add_two_numbers(1,4)
}
