% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celcius_to_fahrenheit.R
\name{celcius_to_fahrenheit}
\alias{celcius_to_fahrenheit}
\title{Convert Celsius to Fahrenheit}
\usage{
celcius_to_fahrenheit(a)
}
\arguments{
\item{a}{(integer or float) Celsius value}
}
\value{
Fahrenheit value
}
\description{
This function converts ...
}
\examples{
celcius_to_fahrenheit(a=25)
}
