% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celcius_to_kelvin.R
\name{celcius_to_kelvin}
\alias{celcius_to_kelvin}
\title{Conversion of Celsius to Kelvin}
\usage{
celcius_to_kelvin(a)
}
\arguments{
\item{a}{(integer or float) Celsius value}
}
\value{
Kelvin value
}
\description{
This function converts ...
}
\examples{
celcius_to_kelvin(a=25)
}
