% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{\code{celsius_to_fahrenheit()}}
\usage{
celsius_to_fahrenheit(tempC)
}
\arguments{
\item{tempC}{(numeric) Temperature in degree Celsius}
}
\value{
Converted temperature (in this case in degree Fahrenheit).
}
\description{
Converts from degrees Celsius to degrees Fahrenheit.
}
\examples{
celsius_to_fahrenheit(tempC = 30)
celsius_to_fahrenheit(tempC = -30)
}
