% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_farenheit.R
\name{celsius_to_farenheit}
\alias{celsius_to_farenheit}
\title{Convert Units from Celsius to Farenheit}
\usage{
celsius_to_farenheit(tempC)
}
\arguments{
\item{tempC}{(numeric) A positive or negative number}
}
\value{
the temperature in Celsius (C°)
}
\description{
The purpose of this function is to convert a temperature in
Celsius (C°) to a temperature in Farenheit (F°).
}
\examples{
celsius_to_farenheit(tempC = 35)
celsius_to_farenheit(35)
}
