% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{\code{celsius_to_kelvin()}}
\usage{
celsius_to_kelvin(tempC)
}
\arguments{
\item{tempC}{(numeric) Temperature in degree Celsius}
}
\value{
Converted temperature (in this case in degree Kelvin).
}
\description{
Converts from degrees Celsius to degrees Kelvin.
}
\examples{
celsius_to_kelvin(tempC = 30)
celsius_to_kelvin(tempC = -30)
}
