% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_to_inches.R
\name{cm_to_inches}
\alias{cm_to_inches}
\title{Convert centimeters to inches}
\usage{
cm_to_inches(cm)
}
\arguments{
\item{cm}{(numeric) A positive number}
}
\value{
Length in inches
}
\description{
Convert centimeters to inches
}
\examples{
cm_to_inches(cm=5)
}
