% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert from one unit type to another}
\usage{
convert(vector, origin, target, print_names = FALSE)
}
\arguments{
\item{vector}{A numeric vector to be converted}

\item{origin}{The catalog symbol of the current unit.
See documentation for list of supportred units.}

\item{target}{The catalog symbol of unit you want to convert to.
See documentation for list of supportred units.}

\item{print_names}{Print the names of the units, useful for debugging.}
}
\description{
\code{convert} accepts a vector of numeric values and
converts them from one unit to another. Approximately 1,300 units
are available.
}
\examples{
convert(1:20, "kg", "g")
convert(1:20, "galUK/min.ft2", "kft/s", print_names = TRUE)
}

