% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_units.R
\name{explore_units}
\alias{explore_units}
\title{Explore conversion table}
\usage{
explore_units()
}
\value{
Launches a Shiny app
}
\description{
A shiny application that lets you take a look at the conversion table.
}
\examples{

}

