% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celcius.R
\name{fahrenheit_to_celcius}
\alias{fahrenheit_to_celcius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celcius(a)
}
\arguments{
\item{a}{(integer or float) Fahrenheit value}
}
\value{
Celsius value
}
\description{
This function converts ...
}
\examples{
fahrenheit_to_celcius(a=345)
}
