% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(temp_F)
}
\arguments{
\item{temp_F}{Temperature in degrees Fahrenheit}
}
\value{
Returns degrees Celsius
}
\description{
The purpose of this function is to convert degrees Fahrenheit into
degrees Celsius. The numbers can be positive or negative but not NA.
}
\examples{
fahrenheit_to_celsius(temp_F = 50)
}
