% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{\code{fahrenheit_to_kelvin()}}
\usage{
fahrenheit_to_kelvin(tempF)
}
\arguments{
\item{tempF}{(numeric) Temperature in degree Fahrenheit}
}
\value{
Converted temperature (in this case in degree Kelvin).
}
\description{
Converts from degrees Fahrenheit to degrees Kelvin.
}
\examples{
fahrenheit_to_kelvin(tempF = 30)
fahrenheit_to_kelvin(tempF = -30)
}
