% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farenheit_to_celsius.R
\name{farenheit_to_celsius}
\alias{farenheit_to_celsius}
\title{Converting Fahrenheit to Celsius}
\usage{
farenheit_to_celsius(tempF)
}
\arguments{
\item{tempF}{}
}
\value{
the temperature in Celsius
}
\description{
Converting Fahrenheit to Celsius
}
\examples{
farenheit_to_celsius(100)
}
