% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farenheit_to_kelvin.R
\name{farenheit_to_kelvin}
\alias{farenheit_to_kelvin}
\title{Converting Fahrenheit to Kelvin}
\usage{
farenheit_to_kelvin(tempF)
}
\arguments{
\item{tempF}{}
}
\value{
Temperature in Kelvin
}
\description{
Converting Fahrenheit to Kelvin
}
\examples{
farenheit_to_kelvin(10)
}
