% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_to_cm.R
\name{ft_to_cm}
\alias{ft_to_cm}
\title{Converting Feet to Centimeters}
\usage{
ft_to_cm(feet)
}
\arguments{
\item{feet}{(numeric)}
}
\value{
Numer in Centimeters
}
\description{
In 1959 the international yard and pound agreement (between the United
States and countries of the Commonwealth of Nations) defined a yard as
being exactly 0.9144 metres, which in turn defined the foot as being
exactly 0.3048 metres (304.8 mm).
}
\examples{
ft_to_cm(3)
}
