% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_conversion_table}
\alias{get_conversion_table}
\title{Return the conversion table
A convenience function to return the lookup table underlying `convert()`}
\usage{
get_conversion_table()
}
\value{
A dataframe of conversion table
}
\description{
Return the conversion table
A convenience function to return the lookup table underlying `convert()`
}
\examples{
a <-  get_conversion_table()
}

