% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_to_cm.R
\name{in_to_cm}
\alias{in_to_cm}
\title{Convert inch to cm}
\usage{
in_to_cm(inch)
}
\arguments{
\item{inch}{(numeric) a positive number}
}
\value{
cm (numeric) a positive number
}
\description{
Convert inch to cm
}
\examples{
in_to_cm(3)
in_to_cm(inch = 3)
}
