% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch_to_cm.R
\name{inch_to_cm}
\alias{inch_to_cm}
\title{Convert inches to centimeters}
\usage{
inch_to_cm(inch)
}
\arguments{
\item{inch}{(numeric) A measurement in inches}
}
\value{
(numeric) The measurement in inches converted to centimeters
}
\description{
This function converts inches to centimeters
}
\examples{
inch_to_cm(inch=12)
}
