% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{is_supported_unit}
\alias{is_supported_unit}
\title{Check whether a unit can be converted.}
\usage{
is_supported_unit(unit)
}
\arguments{
\item{unit}{Name of Unit to be converted}
}
\description{
Check whether a unit can be converted.
}

