% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh_to_knots.R
\name{kmh_to_knots}
\alias{kmh_to_knots}
\title{Kmh to Knots}
\usage{
kmh_to_knots(kmh)
}
\arguments{
\item{kmh}{(numeric) speed in kmh}
}
\value{
The speed in knots
}
\description{
The purpose of this function is to convert a speed from kmh into knots
}
\examples{
kmh_to_knots(50)
}
