% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_opts_to_yaml.R
\name{knitr_opts_to_yaml}
\alias{knitr_opts_to_yaml}
\title{Move Knitr Options from .rmd Setup Chunk into .qmd YAML}
\usage{
knitr_opts_to_yaml(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Input directory for an old .rmd file}

\item{output_dir}{Output directory for a .qmd file with a converted YAML header}
}
\value{
.qmd file
}
\description{
Take an existing .rmd file and convert its YAML header options to be compatible with Quarto. Currently minimal functionality. html output options are automatically set to \code{self-contained: TRUE}. Underscores are replaced with hyphens for all other options.
}
\examples{
 \dontrun{
 knitr_opts_to_yaml(
   input_dir = "path/to/some_an_old_rmd_file.rmd",
   output_dir = "path/to/new_qmd_file.qmd"
  )
}
}
