% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots_to_kmh.R
\name{knots_to_kmh}
\alias{knots_to_kmh}
\title{Knots to Kmh}
\usage{
knots_to_kmh(knots)
}
\arguments{
\item{knots}{(numeric) speed in knots}
}
\value{
The speed in kmh
}
\description{
The purpose of this function is to convert a speed from knots into kmh
}
\examples{
knots_to_kmh(10)
}
