% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miles_to_km.R
\name{miles_to_km}
\alias{miles_to_km}
\title{Convert Miles to Kilometers}
\usage{
miles_to_km(miles)
}
\arguments{
\item{miles}{(numeric) Positive double}
}
\value{
Returns kilometers as a numeric value.
}
\description{
This function takes miles as an input and outputs kilometers.
}
\examples{
miles_to_km(100)
miles_to_km(0.0032)
}
