% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_codechunks.R
\name{process_codechunks}
\alias{process_codechunks}
\title{Process Code Chunks from an R Script}
\usage{
process_codechunks(r_lines_i)
}
\arguments{
\item{r_lines_i}{r_lines file without fences.}
}
\value{
Modified r_lines object tagged with code chunk information
}
\description{
Internal function to add fences to code chunks when converting from R scripts. Used by \code{r_to_qmd()}. Not exported.
}
\keyword{Internal}
