% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_to_r.R
\name{qmd_to_r}
\alias{qmd_to_r}
\title{Convert a Quarto file into an R script}
\usage{
qmd_to_r(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Directory of the qmd file that you want to convert}

\item{output_dir}{Output directory for the new R script}
}
\value{
A .qmd file
}
\description{
Convert a .qmd file into a .R file. Headings are converted into section titles (e.g \verb{# heading -----} as defined by the the shortcut ctrl/cmd + shift + R). Text outside of code chunks is converted into comments. Chunk options defined by \verb{#|} are retained as comments.
}
\examples{
 \dontrun{
 qmd_to_r(
   input_dir = "path/to/some_quarto_file.qmd",
   output_dir = "path/to/new_converted_r_script.R"
  )
}

}
