% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_to_qmd.R
\name{r_to_qmd}
\alias{r_to_qmd}
\title{Convert R Script into a Quarto file}
\usage{
r_to_qmd(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Directory of the R script file that you want to convert}

\item{output_dir}{Output directory for the new qmd file}
}
\value{
An R script file
}
\description{
Convert a .R file into a .qmd file. Section titles are converted into headings (e.g \verb{# heading -----} as defined by the the shortcut ctrl/cmd + shift + R). Comments are converted into text in between codechunks (unless the comments are within the code, in which case they stay inside of the code chunk).
}
\examples{
 \dontrun{
 r_to_qmd(
   input_dir = "path/to/some_R_script.R",
   output_dir = "path/to/new_converted_qmd_file.qmd"
  )
}
}
