% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2cc.R
\name{rad2cc}
\alias{rad2cc}
\title{Cloud cover estimation from radiation data}
\usage{
rad2cc(rsds = NULL, rlds = NULL, rtds = NULL)
}
\arguments{
\item{rsds}{Shortwave downwelling radiation. Ignored if \code{rtds} is provided}

\item{rlds}{Longawe downwelling radiation. Ignored if \code{rtds} is provided.}

\item{rtds}{Total (longwave + shortwave) radiation. If provided, \code{rsds} and \code{rlds} are ignored.}
}
\value{
A climate4R grid of estimated cloud cover
}
\description{
Cloud cover estimation from radiation data
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("rsds.iberia")
data("rlds.iberia")
cc <- rad2cc(rsds.iberia, rlds.iberia)
\dontrun{
require(visualizeR)
spatialPlot(climatology(cc), rev.colors = TRUE, backdrop.theme = "coastline")
}
# Total radiation can be alternatively used:
require(transformeR)
rtds <- gridArithmetics(rsds.iberia, rlds.iberia, operator = "+")
cc2 <- rad2cc(rtds = rtds)
identical(cc2, cc)
}
\seealso{
Other derivation:
\code{\link[=hurs2huss]{hurs2huss()}},
\code{\link[=hurs2tdps]{hurs2tdps()}},
\code{\link[=hurs2w]{hurs2w()}},
\code{\link[=huss2hurs]{huss2hurs()}},
\code{\link[=huss2pvp]{huss2pvp()}},
\code{\link[=ps2psl]{ps2psl()}},
\code{\link[=psl2ps]{psl2ps()}},
\code{\link[=tas2ws]{tas2ws()}},
\code{\link[=tdps2hurs]{tdps2hurs()}}
}
\author{
J. Bedia, S. Herrera
}
\concept{derivation}
