% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducePressure.R
\name{reducePressure}
\alias{reducePressure}
\title{Reduce pressure level}
\usage{
reducePressure(p, t, zgs, direction)
}
\arguments{
\item{p}{A 2D matrix of (sea-level/surface) pressure values (in Pascals)}

\item{t}{A 2D matrix of near-surface temperature (in K)}

\item{zgs}{A matrix of geopotential height (static variable, all rows are equal, in meters)}

\item{paux}{Auxiliar grid (one single member, dropped)}
}
\value{
A 2D matrix with reduced pressure data. Need to be reconverted to a grid using \code{\link[transformeR]{mat2Dto3Darray}} afterwards.
}
\description{
Internal function with formula for presssure reduction
}
\details{
This function is a internal for the derivation functions \code{\link{ps2psl}} and \code{\link{psl2ps}}.
}
\references{
The adjustment formulae of pressure values to other levels are those routinely used by the WMO. These are described in the following technical report:
\url{https://www.wmo.int/pages/prog/www/IMOP/meetings/SI/ET-Stand-1/Doc-10_Pressure-red.pdf}
}
\seealso{
\code{\link{psl2ps}} and \code{\link{ps2psl}}, the exported functions using this one
}
\keyword{internal}
