% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tas2ws.R
\name{tas2ws}
\alias{tas2ws}
\title{Saturation pressure calculation}
\usage{
tas2ws(tas, ps)
}
\arguments{
\item{tas}{Near-surface air temperature}

\item{ps}{Surface pressure}
}
\value{
A climate4R CDM grid of water vapor saturation pressure (in atm)
}
\description{
Calculates the saturation pressure using near-surface air temperature and surface pressure data
}
\details{
The vapour pressure of water is the pressure at which water vapour is in thermodynamic equilibrium
 with its condensed state. At higher pressures water would condense. At this equilibrium condition
 the vapor pressure is the saturation pressure.
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("ps.iberia")
data("tas.iberia")
ws <- tas2ws(tas.iberia, ps.iberia)
\dontrun{
require(visualizeR)
spatialPlot(climatology(ws), backdrop.theme = "coastline")
}
}
\references{
\itemize{
\item Gregory, D. (2000), Atmospheric thermodynamics. By Craig F. Bohren and Bruce A. Albrecht. Oxford University Press. xiv + 402 pp. Price £49.50 (hardback). ISBN 019 5099044. Q.J.R. Meteorol. Soc., 126: 2631-2632. doi:10.1002/qj.49712656815
}
}
\seealso{
Other derivation:
\code{\link[=hurs2huss]{hurs2huss()}},
\code{\link[=hurs2tdps]{hurs2tdps()}},
\code{\link[=hurs2w]{hurs2w()}},
\code{\link[=huss2hurs]{huss2hurs()}},
\code{\link[=huss2pvp]{huss2pvp()}},
\code{\link[=ps2psl]{ps2psl()}},
\code{\link[=psl2ps]{psl2ps()}},
\code{\link[=rad2cc]{rad2cc()}},
\code{\link[=tdps2hurs]{tdps2hurs()}}

Other humidity:
\code{\link[=hurs2huss]{hurs2huss()}},
\code{\link[=hurs2tdps]{hurs2tdps()}},
\code{\link[=hurs2w]{hurs2w()}},
\code{\link[=huss2hurs]{huss2hurs()}},
\code{\link[=huss2pvp]{huss2pvp()}},
\code{\link[=tdps2hurs]{tdps2hurs()}}

Other pressure:
\code{\link[=huss2pvp]{huss2pvp()}},
\code{\link[=ps2psl]{ps2psl()}},
\code{\link[=psl2ps]{psl2ps()}}
}
\author{
J. Bedia, S. Herrera
}
\concept{derivation}
\concept{humidity}
\concept{pressure}
\keyword{internal}
