% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add two numbers together}
\usage{
add_two_numbers(a, b)
}
\arguments{
\item{a}{(integer or float) Number 1 \code{see diesunddas}}

\item{b}{(integer or float) Number 2}
}
\value{
Total of number supplied
}
\description{
User can provide .....
}
\examples{
add_two_number(a=1,b=2)

}
