% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add two numbers together}
\usage{
add_two_numbers(a, b)
}
\arguments{
\item{a}{(numeric) A positive or negative number}

\item{b}{(numeric) A positive or negative number}
}
\value{
The total sum of the two numbers
}
\description{
The purpose of this function is to take two numbers as inputs and add them
together. The numbers' can be positive or negative but not NA.
}
\examples{
add_two_numbers(a = 1, b = 2)
}
