% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Title}
\usage{
celsius_to_fahrenheit(a)
}
\arguments{
\item{a}{denotes the celsius}
}
\value{
the function \code{fahrenheit_to_celsius} returns the fahrenheit
}
\description{
Title
}
\examples{
celsius_to_fahrenheit(a=19)

}
