% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Convert Units from Celsius to Kelvin}
\usage{
celsius_to_kelvin(tempC)
}
\arguments{
\item{tempC}{(numeric) A positive or negative number}
}
\value{
the temperature in Celsius (C°)
}
\description{
The purpose of this function is to convert a temperature in
Celsius (C°) to a temperature in Kelvin (K).
}
\examples{
celsius_to_kelvin(tempC = 0)
celsius_to_kelvin(35)
}
