% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Unit conversion}
\usage{
convert(number, from, to)
}
\arguments{
\item{number}{A numeric}

\item{from}{The standard symbol for the unit being converted from}

\item{to}{The standard symbol for the unit being converted to}
}
\value{
A numeric in the units specified in \code{to}
}
\description{
Unit conversion
}
\examples{
convert(24, 'd', 'h')
convert(1:20, 'lb', 'kg')

}

