% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert from one unit type to another}
\usage{
convert(vector, origin, target)
}
\arguments{
\item{vector}{A numeric vector to be converted}

\item{origin}{The catalog symbol of the current unit.}

\item{target}{The catalog symbol of unit you want to convert to.}
}
\value{
A numeric vector
}
\description{
\code{convert} accepts a vector of numeric values and
converts them from one unit to another. Approximately 1,300 units
are available.
}

