% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert from one unit type to another}
\usage{
convert(vector, origin, target)
}
\arguments{
\item{vector}{A numeric vector to be converted}

\item{origin}{The catalog symbol of the current unit.}

\item{target}{The catalog symbol of unit you want to convert to.}
}
\value{
A numeric vector
}
\description{
Converts numeric vectors from one unit to another Approximately 1500 units
are available.
\code{\link{convert_gadget}} can help you build valid \code{convert()} expressions and
\code{\link{explore_units}} provides an interface for exploring the \code{\link{conversion_table}}
which underlies the computation.
}
\examples{
convert(1:20, "kg", "g")
convert(1:20, "galUK/min.ft2", "kft/s")
\dontrun{
convert(1:20, "kg", "km2")
}

}

