% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{convert_gadget}
\alias{convert_gadget}
\title{Launch interactive conversion gadget}
\usage{
convert_gadget(vector, return_value = FALSE)
}
\arguments{
\item{vector}{A numeric vector to convert. You can change the vector within the app.}

\item{return_value}{By default the gadget generates a \code{convert()} call, if return_value is set to
TRUE it will instead generate a converted vector.}
}
\value{
Insert output text into editor.
}
\description{
It's often difficult to remember the symbol for the units you want to convert.
This gadget shows which units can be converted to one anohter, and a example calculation on sample data.
On exit you can either return a valid \code{convert()} expression or a
converted numeric vector.
}

