% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{convert_gadget}
\alias{convert_gadget}
\title{Launch interactive conversion gadget}
\usage{
convert_gadget(vector)
}
\arguments{
\item{vector}{A numeric vector to convert. You can change the vector within the app.}
}
\value{
Insert text into code with a proper convertr::convert() call.
}
\description{
Launch interactive conversion gadget
}

