% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explore_units.R
\name{explore_units}
\alias{explore_units}
\title{Display the list of supported units}
\usage{
explore_units()
}
\arguments{
\item{shiny}{Whether to launch an interactive shiny app or return a character vector of
unit types.}
}
\description{
\code{convert} accepts a vector of numeric values and
converts them from one unit to another. Approximately 1,300 units
are available.
}
\examples{
explore_units()
}

