% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_units.R
\name{explore_units}
\alias{explore_units}
\title{Explore interactive data table of units, if you want to build a conversion
expression, then it's better to use `convert_gadget()`. This function provides
a fuller view of the units, including some additional metadata.}
\usage{
explore_units()
}
\value{
Launches a Shiny app
}
\description{
Explore interactive data table of units, if you want to build a conversion
expression, then it's better to use `convert_gadget()`. This function provides
a fuller view of the units, including some additional metadata.
}
\examples{

}

