% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(temp_F)
}
\arguments{
\item{temp_F}{(numeric) A positive or negative number}
}
\value{
The temperature in Celsius
}
\description{
The purpose of this function is to take one number as inputs and convert it
from Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(temp_F = 22)
}
