% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{\code{fahrenheit_to_celsius()}}
\usage{
fahrenheit_to_celsius(tempF)
}
\arguments{
\item{tempF}{(numeric) Temperature in degree Fahrenheit}
}
\value{
Converted temperature (in this case in degree Celsius).
}
\description{
Converts from degrees Fahrenheit to degrees Celsius.
}
\examples{
fahrenheit_to_celsius(tempF = 30)
fahrenheit_to_celsius(tempF = -30)
}
