% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{(numeric) Temperature in Fahrenheit}
}
\value{
(numeric) Temperature in Celsius
}
\description{
This function converts Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(32)
}
