% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Convert Fahrenheit to Kelvin}
\usage{
fahrenheit_to_kelvin(a)
}
\arguments{
\item{a}{(integer or float) Fahrenheit value}
}
\value{
Kelvin value
}
\description{
This function converts ...
}
\examples{
fahrenheit_to_kelvin(a=345)
}
