% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_to_cm.R
\name{ft_to_cm}
\alias{ft_to_cm}
\title{Convert Feet to Cm}
\usage{
ft_to_cm(a)
}
\arguments{
\item{a}{(integer or float) Feet value}
}
\value{
Cm value
}
\description{
This function converts ...
}
\examples{
ft_to_cm(a=34)
}
