% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{get_conversion_table}
\alias{get_conversion_table}
\title{Return the conversion table}
\usage{
get_conversion_table()
}
\value{
Conversion table dataframe
}
\description{
A convenience function to return the lookup table underlying \code{convert()}
}
\examples{
a <-  get_conversion_table()
}

