% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurs2huss.R
\name{hurs2huss}
\alias{hurs2huss}
\title{Relative humidity to specific humidity conversion}
\usage{
hurs2huss(hurs, ps, tas)
}
\arguments{
\item{hurs}{Relative humidity grid}

\item{ps}{Surface pressure}

\item{tas}{Near-surface air temperature}
}
\value{
A climate4R CDM grid of specific humidity (in kg/kg)
}
\description{
Relative humidity to specific humidity conversion using surface pressure and temperature
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("ps.iberia")
data("tas.iberia")
data("huss.iberia")
hurs <- huss2hurs(huss = huss.iberia, ps = ps.iberia, tas = tas.iberia)
huss <- hurs2huss(hurs = hurs, ps = ps.iberia, tas = tas.iberia)
identical(huss$Data, huss.iberia$Data)
# Not identical due to rounding errors:
range(huss$Data - huss.iberia$Data)
\dontrun{
require(visualizeR)
spatialPlot(climatology(hurs), rev.colors = TRUE, backdrop.theme = "coastline")
}
}
\references{
\itemize{
\item Gregory, D. (2000), Atmospheric thermodynamics. By Craig F. Bohren and Bruce A. Albrecht. Oxford University Press. xiv + 402 pp. Price £49.50 (hardback). ISBN 019 5099044. Q.J.R. Meteorol. Soc., 126: 2631-2632. doi:10.1002/qj.49712656815
}
}
\seealso{
Other derivation: \code{\link{hurs2w}},
  \code{\link{huss2hurs}}, \code{\link{huss2pvp}},
  \code{\link{ps2psl}}, \code{\link{psl2ps}},
  \code{\link{rad2cc}}, \code{\link{tas2ws}}

Other humidity: \code{\link{hurs2w}},
  \code{\link{huss2hurs}}, \code{\link{huss2pvp}},
  \code{\link{tas2ws}}
}
\author{
J. Bedia, S. Herrera
}
