% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdps2hurs.R
\name{hurs2tdps}
\alias{hurs2tdps}
\title{Relative humidity from Dew-point temperature}
\usage{
hurs2tdps(hurs, tas)
}
\arguments{
\item{hurs}{relative humidity}

\item{tas}{Near-surface air temperature}
}
\value{
A climate4R CDM grid of estimated dew-point temperature (in degC)
}
\description{
Estimate the Relative humidity from Dew-point temperature and air temperaure
}
\note{
The formula is a valid aproximation for moist air (RH>50\%), but can yield very inaccurate results otherwise, so use it with caution.
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\references{
The adjustment formulae of pressure values to other levels are those routinely used by the WMO. These are described in the following technical report:
\url{https://www.wmo.int/pages/prog/www/IMOP/meetings/SI/ET-Stand-1/Doc-10_Pressure-red.pdf}

Lawrence, Mark G., 2005: The relationship between relative humidity and the dewpoint temperature in moist air: A simple conversion and applications. Bull. Amer. Meteor. Soc., 86, 225-233. https://dx.doi.org/10.1175/BAMS-86-2-225
}
\seealso{
tdps2hurs, performing the inverse calculation to derive relative humidity from dew-point temperature and observed temperature

Other derivation:
\code{\link[=hurs2huss]{hurs2huss()}},
\code{\link[=hurs2w]{hurs2w()}},
\code{\link[=huss2hurs]{huss2hurs()}},
\code{\link[=huss2pvp]{huss2pvp()}},
\code{\link[=ps2psl]{ps2psl()}},
\code{\link[=psl2ps]{psl2ps()}},
\code{\link[=rad2cc]{rad2cc()}},
\code{\link[=tas2ws]{tas2ws()}},
\code{\link[=tdps2hurs]{tdps2hurs()}}

Other humidity:
\code{\link[=hurs2huss]{hurs2huss()}},
\code{\link[=hurs2w]{hurs2w()}},
\code{\link[=huss2hurs]{huss2hurs()}},
\code{\link[=huss2pvp]{huss2pvp()}},
\code{\link[=tas2ws]{tas2ws()}},
\code{\link[=tdps2hurs]{tdps2hurs()}}
}
\author{
J. Bedia
}
\concept{derivation}
\concept{humidity}
