% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurs2w.R
\name{hurs2w}
\alias{hurs2w}
\title{Mixing (mass) ratio of water vapor}
\usage{
hurs2w(ps, tas, hurs)
}
\arguments{
\item{ps}{Surface pressure}

\item{tas}{Near-surface air temperature}

\item{hurs}{Relative humidity grid}
}
\value{
A climate4R CDM grid
}
\description{
Calculate the mass ratio of water vapor in air from relative humidity, pressure and temperature
}
\details{
The mixing ratio usually refers to the mass ratio (\eqn{\zeta_i}), which is defined as the mass of a constituent \eqn{m_i}
divided by the total mass of all other constituents in a mixture:

\deqn{\zeta_{i}=\frac{m_i}{m_{tot}-m_{i}}}

The mass ratio of water vapor in air can be used to describe humidity (\url{https://en.wikipedia.org/wiki/Mixing_ratiohttps://en.wikipedia.org/wiki/Mixing_ratio}).
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("ps.iberia")
data("tas.iberia")
data("huss.iberia")
# First, hurs is derived from huss:
hurs <- huss2hurs(huss.iberia, ps.iberia, tas.iberia)
w <- hurs2w(ps.iberia, tas.iberia, hurs)
\dontrun{
require(visualizeR)
spatialPlot(climatology(w), rev.colors = TRUE, backdrop.theme = "coastline")
}
}
\references{
\itemize{
\item Gregory, D. (2000), Atmospheric thermodynamics. By Craig F. Bohren and Bruce A. Albrecht. Oxford University Press. xiv + 402 pp. Price £49.50 (hardback). ISBN 019 5099044. Q.J.R. Meteorol. Soc., 126: 2631-2632. doi:10.1002/qj.49712656815
}

\itemize{
\item Gregory, D. (2000), Atmospheric thermodynamics. By Craig F. Bohren and Bruce A. Albrecht. Oxford University Press. xiv + 402 pp. Price £49.50 (hardback). ISBN 019 5099044. Q.J.R. Meteorol. Soc., 126: 2631-2632. doi:10.1002/qj.49712656815
}
}
\seealso{
Other derivation: \code{\link{hurs2huss}},
  \code{\link{huss2hurs}}, \code{\link{huss2pvp}},
  \code{\link{ps2psl}}, \code{\link{psl2ps}},
  \code{\link{rad2cc}}, \code{\link{tas2ws}},
  \code{\link{tdps2hurs}}

Other humidity: \code{\link{hurs2huss}},
  \code{\link{huss2hurs}}, \code{\link{huss2pvp}},
  \code{\link{tas2ws}}, \code{\link{tdps2hurs}}
}
\author{
J. Bedia, S. Herrera
}
\concept{derivation}
\concept{humidity}
\keyword{internal}
