% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huss2hurs.R
\name{huss2hurs}
\alias{huss2hurs}
\title{Specific humidity to relative humidity conversion}
\usage{
huss2hurs(huss, ps, tas)
}
\arguments{
\item{huss}{Specific humidity grid}

\item{ps}{Surface pressure}

\item{tas}{Near-surface air temperature}
}
\value{
A climate4R grid
}
\description{
Specific humidity to relative humidity conversion using surface pressure and temperature
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("ps.iberia")
data("tas.iberia")
data("huss.iberia")
hurs <- huss2hurs(huss.iberia, ps.iberia, tas.iberia)
\dontrun{
require(visualizeR)
spatialPlot(climatology(hurs), rev.colors = TRUE, backdrop.theme = "coastline")
}
}
\seealso{
Other derivation: \code{\link{rad2cc}},
  \code{\link{tas2ws}}
}
\author{
J. Bedia, S. Herrera
}
