% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huss2pvp.R
\name{huss2pvp}
\alias{huss2pvp}
\title{Specific humidity to partial vapor pressure conversion}
\usage{
huss2pvp(huss, ps)
}
\arguments{
\item{huss}{Specific humidity grid}

\item{ps}{Surface pressure}
}
\value{
A climate4R CDM grid of partial vapor pressure (in Pascals)
}
\description{
Specific humidity to partial vapor pressure conversion using surface pressure and specific humidity
}
\details{
The partial pressure of a gaseous constituent of air (water vapor in this case)
 is the pressure which it alone would exert with unchanged temperature and number of moles per unit volume.
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("ps.iberia")
data("huss.iberia")
pvp <- huss2pvp(huss = huss.iberia, ps = ps.iberia)
\dontrun{
require(visualizeR)
spatialPlot(climatology(pvp), rev.colors = TRUE, backdrop.theme = "coastline")
}
}
\references{
\itemize{
\item Gregory, D. (2000), Atmospheric thermodynamics. By Craig F. Bohren and Bruce A. Albrecht. Oxford University Press. xiv + 402 pp. Price £49.50 (hardback). ISBN 019 5099044. Q.J.R. Meteorol. Soc., 126: 2631-2632. doi:10.1002/qj.49712656815
}
}
\seealso{
Other derivation: \code{\link{hurs2huss}},
  \code{\link{hurs2tdps}}, \code{\link{hurs2w}},
  \code{\link{huss2hurs}}, \code{\link{ps2psl}},
  \code{\link{psl2ps}}, \code{\link{rad2cc}},
  \code{\link{tas2ws}}, \code{\link{tdps2hurs}}

Other humidity: \code{\link{hurs2huss}},
  \code{\link{hurs2tdps}}, \code{\link{hurs2w}},
  \code{\link{huss2hurs}}, \code{\link{tas2ws}},
  \code{\link{tdps2hurs}}

Other pressure: \code{\link{ps2psl}}, \code{\link{psl2ps}},
  \code{\link{tas2ws}}
}
\author{
J. Bedia, S. Herrera
}
\concept{derivation}
\concept{humidity}
\concept{pressure}
