% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_to_cm.R
\name{in_to_cm}
\alias{in_to_cm}
\title{Convert inches to centimeters}
\usage{
in_to_cm(inches)
}
\arguments{
\item{inches}{(numeric) A positive number}
}
\value{
The length in centimeters
}
\description{
The purpose of this function is to take one number as input and convert it
from inches to centimeters.
}
\examples{
in_to_cm(inches = 5)
}
