% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_to_cm.R
\name{in_to_cm}
\alias{in_to_cm}
\title{Converting Inches to Centimeters}
\usage{
in_to_cm(inch)
}
\arguments{
\item{inch}{(numeric) A positive or negative number}
}
\value{
number in Centimeter
}
\description{
Converting Inches to Centimeters
}
\examples{
in_to_cm(3)
}
