% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps2psl.R
\name{ps2psl}
\alias{ps2psl}
\title{Surface pressure reduction to sea-level}
\usage{
ps2psl(ps, tas, zgs)
}
\arguments{
\item{ps}{Surface pressure}

\item{tas}{Near-surface air temperature}

\item{zgs}{surface geopotential height. This is a static variable (i.e., constant in time).
Thus, member and time dimensions must be either singletons or just dropped.}
}
\value{
A climate4R CDM grid of estimated sea-level pressure (in Pascals)
}
\description{
Estimate the sea-level pressure from surface pressure, using temperature and surface geopotential height
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("ps.iberia")
data("tas.iberia")
data("zgs.iberia")
psl.hat <- ps2psl(ps = ps.iberia, tas = tas.iberia, zgs = zgs.iberia)
require(transformeR)
# Error map:
psl.err <- gridArithmetics(psl.hat, psl.iberia, operator = "-")
\dontrun{
require(visualizeR)
err <- climatology(psl.err)
spatialPlot(err, set.min = -10000,
            at = seq(-10000, 10000, 500),
            color.theme = "Spectral", rev.colors = TRUE,
            backdrop.theme = "coastline",
            main = "Mean estimate error (Pascals)")
}
}
\references{
The adjustment formulae of pressure values to other levels are those routinely used by the WMO. These are described in the following technical report:
\url{https://www.wmo.int/pages/prog/www/IMOP/meetings/SI/ET-Stand-1/Doc-10_Pressure-red.pdf}
}
\seealso{
Other derivation: \code{\link{hurs2huss}},
  \code{\link{hurs2w}}, \code{\link{huss2hurs}},
  \code{\link{huss2pvp}}, \code{\link{psl2ps}},
  \code{\link{rad2cc}}, \code{\link{tas2ws}}

Other pressure: \code{\link{huss2pvp}},
  \code{\link{psl2ps}}, \code{\link{tas2ws}}
}
\author{
J. Bedia, S. Herrera
}
