% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psl2ps.R
\name{psl2ps}
\alias{psl2ps}
\title{Sea-level pressure to surface pressure}
\usage{
psl2ps(psl, tas, zgs)
}
\arguments{
\item{psl}{Sea-surface pressure}

\item{tas}{Near-surface air temperature}

\item{zgs}{surface geopotential height. This is a static variable (i.e., constant in time).
Thus, member and time dimensions must be either singletons or just dropped.}
}
\value{
A climate4R CDM grid of estimated surface pressure (in Pascals)
}
\description{
Estimate the surface pressure from sea-level pressure, using temperature and surface geopotential height
}
\details{
Surface pressure is often needed for other variable derivations, but in many models is often not available, while mean sea-level
pressure is usually provided.
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
data("psl.iberia")
data("tas.iberia")
data("zgs.iberia")
ps.hat <- psl2ps(psl = psl.iberia, tas = tas.iberia, zgs = zgs.iberia)
require(transformeR)
# Error map:
data("ps.iberia")
ps.err <- gridArithmetics(ps.hat, ps.iberia, operator = "-")
\dontrun{
require(visualizeR)
err <- climatology(ps.err)
spatialPlot(err, set.max = 10000,
            at = seq(-10000, 10000, 500),
            color.theme = "Spectral", rev.colors = TRUE,
            backdrop.theme = "coastline",
            main = "Mean estimate error (Pascals)")
}
}
\references{
The adjustment formulae of pressure values to other levels are those routinely used by the WMO. These are described in the following technical report:
\url{https://www.wmo.int/pages/prog/www/IMOP/meetings/SI/ET-Stand-1/Doc-10_Pressure-red.pdf}
}
\seealso{
Other derivation: 
\code{\link{hurs2huss}()},
\code{\link{hurs2tdps}()},
\code{\link{hurs2w}()},
\code{\link{huss2hurs}()},
\code{\link{huss2pvp}()},
\code{\link{ps2psl}()},
\code{\link{rad2cc}()},
\code{\link{tas2ws}()},
\code{\link{tdps2hurs}()}

Other pressure: 
\code{\link{huss2pvp}()},
\code{\link{ps2psl}()},
\code{\link{tas2ws}()}
}
\author{
J. Bedia, S. Herrera
}
\concept{derivation}
\concept{pressure}
