% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2cc.R
\name{rad2cc}
\alias{rad2cc}
\title{Cloud cover estimation from radiation data}
\usage{
rad2cc(rsds = NULL, rlds = NULL, rtds = NULL)
}
\arguments{
\item{rsds}{Shortwave downwelling radiation. Ignored if \code{rtds} is provided}

\item{rlds}{Longawe downwelling radiation. Ignored if \code{rtds} is provided.}

\item{rtds}{Total (longwave + shortwave) radiation. If provided, \code{rsds} and \code{rlds} are ignored.}
}
\value{
A climate4R grid of estimated cloud cover
}
\description{
Cloud cover estimation from radiation data
}
\examples{
data("rsds.iberia")
data("rlds.iberia")
cc <- rad2cc(rsds.iberia, rlds.iberia)
library(visualizeR)
spatialPlot(climatology(cc), rev.colors = TRUE, backdrop.theme = "coastline")
# Total radiation can be alternatively used:
library(transformeR)
rtds <- gridArithmetics(rsds.iberia, rlds.iberia, operator = "+")
cc2 <- rad2cc(rtds = rtds)
identical(cc2, cc)
}
\author{
J. Bedia, S. Herrera
}
