% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tas2ws.R
\name{tas2ws}
\alias{tas2ws}
\title{Saturation pressure calculation}
\usage{
tas2ws(tas, ps)
}
\arguments{
\item{tas}{Near-surface air temperature}

\item{ps}{Surface pressure}
}
\value{
A climate4R grid of saturation pressure (in Pascals)
}
\description{
Calculates the saturation pressure using near-surface air temperature and surface pressure data
}
\section{Input units}{

In principle, the derivation functions will work regardless of the input units, as all units are internally
 converted as necessary according to the specific formulae definitions.
  However, the units string (see \code{\link[transformeR]{getGridUnits}}) must be parseable. In the same vein,
  the input units need to be convertible to the required ones. Unit consistency is internally achieved
  by the function \code{\link{udConvertGrid}}.
}

\examples{
\dontrun{
data("ps.iberia")
data("tas.iberia")
ws <- tas2ws(tas.iberia, ps.iberia)
library(visualizeR)
spatialPlot(climatology(ws), backdrop.theme = "coastline")}
}
\seealso{
Other derivation: \code{\link{huss2hurs}},
  \code{\link{rad2cc}}
}
\author{
J. Bedia, S. Herrera
}
\keyword{internal}
